Added the minimum and maximum frequency parameters for PL source. These are either read in from input or taken from c2ray_parameters.
Added a field photo_out to the structure photrates. Needed for collecting
outgoing photons for photon loss calculation. Note that this field
also was added to the two routines working on photrates structures:
photrates_add and set_photrates_to_zero

Fixed (currently harmless) bug: outgoing heating rate is in-cell but
was calculated as in+cell.