Update of photon_loss_src (photon loss per source) calculation. It was originally assumed we would track this for each frequency subbin, but this is not actually the case.
The use of a frequency dependent photon loss did interfer with
the checking of losses from subboxes as the code never used
more than one subbox.

Changed the photon_loss_src variable from array to scalar and
testing it now with subbox criterion. For the standard test
problem we seem to loose about 0.5% from the box. This implies
that we can use subboxes to test for loss of photons, because
it is not a very small number.