This is a new version of sourceprops_cubep3m.F90 which is using preprocessor options to deal with the presence of different types of sources: PL - power law sources QUASARS - quasar sources Previously there were no quasar sources and the PL sources were always assumed to be there (but could all be zero).
This change (initiated by Hannah Ross and Ilian Iliev) affects other
files such as the different radiation modules. However, these have not
yet been updated yet.

This means that in order to compile the code, it is required to use
the -DPL option. All entries for C2Ray_3D_cubep3m in the Makefile
have this option added to it.