Changes triggered by a restructuring (factorization) of the module radiation_sed_parameters.f90. This module now contains more subroutines and is hopefully easier to understand.
In order to make the variable names consistent, the variables
S_star and S_star_nominal are now known as bb_S_star and bb_S_star_nominal.
This impacts several modules which use these variables.

This commit compiles for the test case without the -DPL and -DQUASAR
compile options.