Changes to radiation_sed_parameters after debugging.
It already worked when parameters were taken from the nominal values
in sed_parameters but black body failed when parameters came from
command line since h/kT was not initialized.

Added proper limits when calculating the photon rate / luminosity
in the three bands. In the previous version the frequency limits used
in the code were ignored in this diagnostic output.